split out agentsmd scaffold skill
