+++ docs/skill-registry.md
@@ -2,10 +2,11 @@
 
 <!-- Generated by scripts/validate_skills.py --write. Do not edit manually. -->
 
-Total installable skills: 86
+Total installable skills: 87
 
 | Name | Category | Format | Lang | Runtime | Tags | Path | Description |
 |---|---|---|---|---|---|---|---|
+| `agentsmd-scaffold` | AI & Agent Workflow | directory | en | unspecified | agent | [skills/agentsmd-scaffold/SKILL.md](../skills/agentsmd-scaffold/SKILL.md) | Generate or update repository-specific AGENTS.md instruction files from real repo evidence. Use when asked to create, design, scaffold, split, or improve root or scoped AGENTS.md files for Codex/Claude/agent workflows, especially when a repo needs directory-specific rules, validation commands, generated-file boundaries, or a short agent onboarding router. |
 | `api-design` | API & Backend | directory | en | unspecified | backend | [skills/api-design/SKILL.md](../skills/api-design/SKILL.md) | REST/GraphQL/gRPC API design best practices. Use when designing APIs, defining contracts, handling versioning. Covers OpenAPI 3.2, GraphQL Federation, gRPC streaming. |
 | `app-ui-design` | UI/UX & Frontend | directory | en | unspecified | frontend, ux | [skills/app-ui-design/SKILL.md](../skills/app-ui-design/SKILL.md) | Mobile app UI design expert for iOS and Android. Use when designing app interfaces, creating design systems, ensuring accessibility, or following platform guidelines. Covers Material Design 3, Human Interface Guidelines, color theory, typography, and 2025 trends. |
 | `architecture-foundation` | Development Architecture | directory | en | unspecified | architecture, go, refactoring, rust | [skills/architecture-foundation/SKILL.md](../skills/architecture-foundation/SKILL.md) | Design architecture foundations before implementation. Use when asked to design or refactor architecture, choose Rust/Go crate, package, module, runtime, workflow, or service boundaries, compare mature project architecture, prevent stacked one-off PRs, audit migration debt in existing repos, or produce a spec before coding. |
@@ -62,7 +63,7 @@ Total installable skills: 86
 | `python-project` | Development Architecture | directory | en | unspecified | architecture, python | [skills/python-project/SKILL.md](../skills/python-project/SKILL.md) | Modern Python project architecture guide for 2025. Use when creating Python projects (APIs, CLI, data pipelines). Covers uv, Ruff, Pydantic, FastAPI, and async patterns. |
 | `react-best-practices` | Development Architecture | directory | en | unspecified | javascript, performance, react | [skills/react-best-practices/SKILL.md](../skills/react-best-practices/SKILL.md) | React and Next.js performance playbook distilled from Vercel Engineering guidance. Use when building, reviewing, or refactoring React/Next.js code for waterfalls, bundle size, rendering cost, and client/server data flow. |
 | `react-hooks-best-practices` | Development Architecture | directory | en | unspecified | react | [skills/react-hooks-best-practices/SKILL.md](../skills/react-hooks-best-practices/SKILL.md) | Provides React patterns for hooks, effects, refs, and component design. Covers escape hatches, anti-patterns, and correct effect usage. Use when reading or writing React components with hooks or effects. |
-| `repo-agent-context-audit` | AI & Agent Workflow | directory | en | unspecified | agent, code-review | [skills/repo-agent-context-audit/SKILL.md](../skills/repo-agent-context-audit/SKILL.md) | Audit, design, and scaffold a repository's agent-readable context, including root and scoped AGENTS.md files, CLAUDE.md, WARP.md, CONTRIBUTING.md, .agents/skills, and specs/ PRODUCT.md and TECH.md contracts. Use when asked to review, generate, create, standardize, or improve repo instructions, agent onboarding, scoped AGENTS.md rules, spec workflows, or cross-repo agent-context conventions. |
+| `repo-agent-context-audit` | AI & Agent Workflow | directory | en | unspecified | agent, code-review | [skills/repo-agent-context-audit/SKILL.md](../skills/repo-agent-context-audit/SKILL.md) | Audit and recommend improvements for a repository's agent-readable context, including AGENTS.md, CLAUDE.md, WARP.md, CONTRIBUTING.md, .agents/skills, and specs/ PRODUCT.md and TECH.md contracts. Use when asked to review, score, assess, or standardize repo instructions, agent onboarding, spec workflows, or cross-repo agent-context conventions. Use agentsmd-scaffold instead when the user wants to generate or apply root/scoped AGENTS.md files. |
 | `rust-best-practices` | Development Architecture | directory | en | unspecified | backend, performance, rust | [skills/rust-best-practices/SKILL.md](../skills/rust-best-practices/SKILL.md) | Use when writing, modifying, or reviewing Rust code. ALWAYS invoke before Rust edits; covers Microsoft Pragmatic Rust guidance for error handling, API design, performance, and idiomatic patterns. |
 | `rust-project` | Development Architecture | directory | en | unspecified | architecture, rust | [skills/rust-project/SKILL.md](../skills/rust-project/SKILL.md) | Modern Rust project architecture guide for 2025. Use when creating Rust projects (CLI, web services, libraries). Covers workspace structure, error handling, async patterns, and idiomatic Rust best practices. |
 | `rustdesk-doctor` | Operations & Deploy | directory | zh | unspecified | debugging, system | [skills/rustdesk-doctor/SKILL.md](../skills/rustdesk-doctor/SKILL.md) | 诊断 RustDesk 连接问题。当用户说 RustDesk 连不上、无法远程、ID 不显示、注册失败时使用 |
