+++ registry/skills.json
@@ -1,5 +1,21 @@
 [
   {
+    "name": "agentsmd-scaffold",
+    "category": "AI & Agent Workflow",
+    "format": "directory",
+    "path": "skills/agentsmd-scaffold/SKILL.md",
+    "description": "Generate or update repository-specific AGENTS.md instruction files from real repo evidence. Use when asked to create, design, scaffold, split, or improve root or scoped AGENTS.md files for Codex/Claude/agent workflows, especially when a repo needs directory-specific rules, validation commands, generated-file boundaries, or a short agent onboarding router.",
+    "language": "en",
+    "compatibility": {
+      "runtimes": [
+        "unspecified"
+      ]
+    },
+    "tags": [
+      "agent"
+    ]
+  },
+  {
     "name": "api-design",
     "category": "API & Backend",
     "format": "directory",
@@ -964,7 +980,7 @@
     "category": "AI & Agent Workflow",
     "format": "directory",
     "path": "skills/repo-agent-context-audit/SKILL.md",
-    "description": "Audit, design, and scaffold a repository's agent-readable context, including root and scoped AGENTS.md files, CLAUDE.md, WARP.md, CONTRIBUTING.md, .agents/skills, and specs/ PRODUCT.md and TECH.md contracts. Use when asked to review, generate, create, standardize, or improve repo instructions, agent onboarding, scoped AGENTS.md rules, spec workflows, or cross-repo agent-context conventions.",
+    "description": "Audit and recommend improvements for a repository's agent-readable context, including AGENTS.md, CLAUDE.md, WARP.md, CONTRIBUTING.md, .agents/skills, and specs/ PRODUCT.md and TECH.md contracts. Use when asked to review, score, assess, or standardize repo instructions, agent onboarding, spec workflows, or cross-repo agent-context conventions. Use agentsmd-scaffold instead when the user wants to generate or apply root/scoped AGENTS.md files.",
     "language": "en",
     "compatibility": {
       "runtimes": [
