+++ skills/agentsmd-scaffold/references/scaffold-agents.md
